refactor(RaceController.js): make class to file
